Reorganize `astropy.units.physical`
The implementation of the physical types machinery is now in a separate
file from the definition of `PhysicalType` instances. This will allow
adding a stub file for the `PhysicalType` instances while keeping the
inline annotations for the `PhysicalType` class and related functions.
